improve thread management

+++ resources/lib/downloads.py
@@ -38,8 +38,29 @@ class Logger():
 
 class Downloads:
 
-    def __init__(self):
-        return
+    def __init__(self, cleanup=False):
+        self.pending = []
+        self.thread = []
+        self.process = []
+        self.alive = True
+        # ディレクトリのクリーンアップ
+        if cleanup:
+            # jsonファイル
+            for json_file in glob.glob(os.path.join(Const.DOWNLOAD_PATH, '*.json')):
+                mp3_file = re.sub(r'\.json$', '.mp3', json_file)
+                # 対応するmp3ファイルがない場合は削除
+                if not os.path.isfile(mp3_file): os.remove(json_file)
+            # 一時ファイル
+            for tmp_file in glob.glob(os.path.join(Const.DOWNLOAD_PATH, '.*.mp3')):
+                os.remove(tmp_file)
+            # キューファイル
+            if os.path.isfile(Const.QUEUE_FILE):
+                os.remove(Const.QUEUE_FILE)
+
+    def abort(self):
+        self.alive = False
+        for p in self.process:
+            if p.returncode is None: p.kill()
 
     def status(self, id, ft):
         # 番組ID
@@ -55,12 +76,72 @@ class Downloads:
         if os.path.isfile(mp3_file): status += 4
         return status
 
-    def add(self, id, name, ft, to, title, description, source, delay, key=''):
+    def enqueue(self, id, name, ft, to, title, description, source, delay):
         # 既存の番組情報ファイルの有無をチェック
         status = self.status(id, ft)
         if status & 4: return 'Already downloaded'
         if status & 2: return 'Now downloading'
         if status & 1: return 'Already scheduled'
+        # 既存のキューをファイルから読み込む
+        queue = []
+        if os.path.isfile(Const.QUEUE_FILE):
+            queue = read_json(Const.QUEUE_FILE)
+        # キューに追加
+        queue.append({
+            'id': id,
+            'name': name,
+            'ft': ft,
+            'to': to,
+            'title': title,
+            'description': description,
+            'source': source,
+            'delay': delay
+        })
+        # キューをファイルにに書き込む
+        write_json(Const.QUEUE_FILE, queue)
+
+    def filter(self):
+        remaining = []
+        now = datetime.datetime.now()
+        for m in self.pending:
+            p = m['program']
+            k = m['keyword']
+            # 開始直前であれば保存処理を開始
+            start = strptime(p['ft'], '%Y%m%d%H%M%S') - now
+            end = strptime(p['to'], '%Y%m%d%H%M%S') - now
+            # すでに終了している番組
+            if end.days < 0:
+                # 次の番組のダウンロード処理へ
+                continue
+            # すでに開始している番組、開始していないがConst.PREP_INTERVAL以内に開始する番組
+            elif start.days < 0 or (start.days == 0 and start.seconds < Const.PREP_INTERVAL):
+                # ダウンロード実行
+                status = self.start(
+                    id=p['id'],
+                    name=p['name'],
+                    ft=p['ft'],
+                    to=p['to'],
+                    title=p['title'],
+                    description=p['description'],
+                    source=p['source'],
+                    delay=p['delay'],
+                    key=k['key'])
+                # 正常終了したらログに書き出す
+                if status is None:
+                    log('download scheduled. id:{id}, start:{start}, title:{title}, keyword:{keyword}'.format(
+                        id = p['id'],
+                        start = p['ft'],
+                        title = p['title'],
+                        keyword = k['key']))
+                # 次の番組のダウンロード処理へ
+                continue
+            # Const.PREP_INTERVAL以降に開始する番組
+            else:
+                # ダウンロード待ちの番組としてリストに追加
+                remaining.append(m)
+        self.pending = remaining
+
+    def start(self, id, name, ft, to, title, description, source, delay, key=''):
         # 番組ID
         gtvid = '%s_%s' % (id, ft);
         # ファイルパス
@@ -168,7 +249,9 @@ class Downloads:
             'tmp_file': tmp_file,
             'mp3_file': mp3_file
         }
-        threading.Thread(target=self.__thread, args=[data]).start()
+        t = threading.Thread(target=self.__thread, args=[data])
+        t.start()
+        self.thread.append(t)
 
     def __thread(self, data):
         # コマンドライン
@@ -195,36 +278,44 @@ class Downloads:
                 tit3=data['tit3'],
                 tmp_file=data['tmp_file'])
         # 開始待機
-        time.sleep(data['wait'])
-        # 開始通知
-        notify('Download started "{title}"'.format(title=data['title']))
-        # ログ書き込み初期化
-        logger = Logger(Params.LOG_FILE)
-        # ダウンロード開始
-        p = subprocess.Popen(command, stderr=logger.handle, stdout=logger.handle, shell=True)
-        # ログ
-        logger.write('[{pid}] Download started. title:"{title}" file:"{mp3_file}"'.format(pid=p.pid, title=data['title'], mp3_file=data['mp3_file']))
-        # ダウンロード終了を待つ
-        p.wait()
-        # ダウンロード結果に応じて後処理
-        if p.returncode:
-            # 失敗したときはjsonファイル、一時ファイルを削除
+        t = data['wait']
+        while t > 0 and self.alive:
+            time.sleep(1)
+            t -= 1
+        if self.alive:
+            # 開始通知
+            notify('Download started "{title}"'.format(title=data['title']))
+            # ログ書き込み初期化
+            logger = Logger(Params.LOG_FILE)
+            # ダウンロード開始
+            p = subprocess.Popen(command, stderr=logger.handle, stdout=logger.handle, shell=True)
+            self.process.append(p)
+            # ログ
+            logger.write('[{pid}] Download started. title:"{title}" file:"{mp3_file}"'.format(pid=p.pid, title=data['title'], mp3_file=data['mp3_file']))
+            # ダウンロード終了を待つ
+            p.wait()
+            # ダウンロード結果に応じて後処理
+            if p.returncode:
+                # 失敗したときはjsonファイル、一時ファイルを削除
+                os.remove(data['json_file'])
+                os.remove(data['tmp_file'])
+            else:
+                # 一時ファイルをリネーム
+                os.rename(data['tmp_file'], data['mp3_file'])
+                # rssファイル生成
+                RSS().create()
+            # 完了通知
+            if p.returncode:
+                notify('Download failed "{title}"'.format(title=data['title']), error=True)
+                # ログ
+                logger.write('[{pid}] Download failed. returncode:{returncode}'.format(pid=p.pid, returncode=p.returncode))
+            else:
+                notify('Download completed "{title}"'.format(title=data['title']))
+                # ログ
+                logger.write('[{pid}] Download completed.'.format(pid=p.pid))
+        else:
+            # 中断した時はjsonファイルを削除
             os.remove(data['json_file'])
-            os.remove(data['tmp_file'])
-        else:
-            # 一時ファイルをリネーム
-            os.rename(data['tmp_file'], data['mp3_file'])
-            # rssファイル生成
-            RSS().create()
-        # 完了通知
-        if p.returncode:
-            notify('Download failed "{title}"'.format(title=data['title']), error=True)
-            # ログ
-            logger.write('[{pid}] Download failed. returncode:{returncode}'.format(pid=p.pid, returncode=p.returncode))
-        else:
-            notify('Download completed "{title}"'.format(title=data['title']))
-            # ログ
-            logger.write('[{pid}] Download completed.'.format(pid=p.pid))
 
     def delete(self, gtvid=None, key=None):
         # ファイル削除
@@ -256,9 +347,6 @@ class Downloads:
             if os.path.isfile(mp3_file):
                 plist.append(read_json(json_file))
             else:
-                # ダウンロード中の場合もあるので削除はしない
-                #os.remove(json_file)
-                #if os.path.isfile(tmp_file): os.remove(tmp_file)
                 log('file lost (or in downloading):{file}'.format(file=mp3_file))
         # 日付フォーマット
         h = Holiday()
